requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.5
    # via
    #   aiohttp-retry
    #   langchain-community
    #   pinecone
aiohttp-retry==2.9.1
    # via pinecone
aiosignal==1.4.0
    # via aiohttp
altair==6.1.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   streamlit
attrs==26.1.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==7.1.1
    # via streamlit
certifi==2026.4.22
    # via
    #   httpcore
    #   httpx
    #   pinecone
    #   requests
charset-normalizer==3.4.7
    # via requests
click==8.3.3
    # via
    #   streamlit
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
dataclasses-json==0.6.7
    # via langchain-community
distro==1.9.0
    # via openai
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
gitdb==4.0.12
    # via gitpython
gitpython==3.1.49
    # via streamlit
greenlet==3.5.0
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via streamlit
httpx==0.28.1
    # via
    #   langchain-pinecone
    #   langgraph-sdk
    #   langsmith
    #   openai
httpx-sse==0.4.3
    # via langchain-community
idna==3.13
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
itsdangerous==2.2.0
    # via streamlit
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jiter==0.14.0
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.1.1
    # via jsonpatch
jsonschema==4.26.0
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
langchain==1.2.17
    # via -r requirements.in
langchain-classic==1.0.5
    # via langchain-community
langchain-community==0.4.1
    # via -r requirements.in
langchain-core==1.3.2
    # via
    #   langchain
    #   langchain-classic
    #   langchain-community
    #   langchain-openai
    #   langchain-pinecone
    #   langchain-text-splitters
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-openai==1.2.1
    # via
    #   -r requirements.in
    #   langchain-pinecone
langchain-pinecone==0.2.13
    # via -r requirements.in
langchain-protocol==0.0.15
    # via langchain-core
langchain-text-splitters==1.1.2
    # via langchain-classic
langchainhub==0.1.21
    # via -r requirements.in
langgraph==1.1.10
    # via langchain
langgraph-checkpoint==4.0.3
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==1.0.13
    # via langgraph
langgraph-sdk==0.3.13
    # via langgraph
langsmith==0.8.0
    # via
    #   langchain-classic
    #   langchain-community
    #   langchain-core
markupsafe==3.0.3
    # via jinja2
marshmallow==3.26.2
    # via dataclasses-json
multidict==6.7.1
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via typing-inspect
narwhals==2.20.0
    # via altair
numpy==2.4.4
    # via
    #   langchain-community
    #   langchain-pinecone
    #   pandas
    #   pydeck
    #   streamlit
openai==2.33.0
    # via langchain-openai
orjson==3.11.8
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.12.2
    # via langgraph-checkpoint
packaging==24.2
    # via
    #   altair
    #   langchain-core
    #   langchainhub
    #   langsmith
    #   marshmallow
    #   pinecone-plugin-assistant
    #   streamlit
pandas==3.0.2
    # via streamlit
pillow==12.2.0
    # via streamlit
pinecone[asyncio]==7.3.0
    # via langchain-pinecone
pinecone-plugin-assistant==1.8.0
    # via pinecone
pinecone-plugin-interface==0.0.7
    # via pinecone
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
protobuf==7.34.1
    # via streamlit
pyarrow==24.0.0
    # via streamlit
pydantic==2.13.3
    # via
    #   langchain
    #   langchain-classic
    #   langchain-core
    #   langchain-pinecone
    #   langgraph
    #   langsmith
    #   openai
    #   pydantic-settings
pydantic-core==2.46.3
    # via pydantic
pydantic-settings==2.14.0
    # via langchain-community
pydeck==0.9.2
    # via streamlit
python-dateutil==2.9.0.post0
    # via
    #   pandas
    #   pinecone
python-dotenv==1.2.2
    # via
    #   -r requirements.in
    #   pydantic-settings
python-multipart==0.0.27
    # via streamlit
pyyaml==6.0.3
    # via
    #   langchain-classic
    #   langchain-community
    #   langchain-core
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2026.4.4
    # via tiktoken
requests==2.33.1
    # via
    #   langchain-classic
    #   langchain-community
    #   langchainhub
    #   langsmith
    #   pinecone-plugin-assistant
    #   requests-toolbelt
    #   streamlit
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
simsimd==6.5.16
    # via langchain-pinecone
six==1.17.0
    # via python-dateutil
smmap==5.0.3
    # via gitdb
sniffio==1.3.1
    # via openai
sqlalchemy==2.0.49
    # via
    #   langchain-classic
    #   langchain-community
starlette==1.0.0
    # via streamlit
streamlit==1.57.0
    # via -r requirements.in
tenacity==9.1.4
    # via
    #   langchain-community
    #   langchain-core
    #   streamlit
tiktoken==0.12.0
    # via langchain-openai
toml==0.10.2
    # via streamlit
tqdm==4.67.3
    # via openai
types-requests==2.33.0.20260503
    # via langchainhub
typing-extensions==4.15.0
    # via
    #   altair
    #   langchain-core
    #   langchain-protocol
    #   openai
    #   pinecone
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   streamlit
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2026.2
    # via pandas
urllib3==2.6.3
    # via
    #   pinecone
    #   requests
    #   types-requests
uuid-utils==0.14.1
    # via
    #   langchain-core
    #   langsmith
uvicorn==0.46.0
    # via streamlit
watchdog==6.0.0
    # via streamlit
websockets==16.0
    # via streamlit
xxhash==3.7.0
    # via
    #   langgraph
    #   langsmith
yarl==1.23.0
    # via aiohttp
zstandard==0.25.0
    # via langsmith
